refactor: 일최고, 최저온도 타입 int -> double

+++ backend/domain-modules/forecast-domain/src/testFixtures/java/com/softeer/domain/DailyTemperatureFixture.java
@@ -11,8 +11,8 @@ public class DailyTemperatureFixture {
     }
 
     public static class DailyTemperatureBuilder {
-        private int highestTemperature = 20;
-        private int lowestTemperature = 10;
+        private double highestTemperature = 20.0;
+        private double lowestTemperature = 10.0;
 
         public DailyTemperatureBuilder highestTemperature(int highestTemperature) {
             this.highestTemperature = highestTemperature;
